pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sls-shared"
version = "1.0.0"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy[asyncio]==2.0.41",
    "psycopg[binary]==3.2.9",
    "PyJWT[crypto]==2.9.0",
    "fastapi==0.115.12",
    "pydantic==2.11.4",
    "pydantic-settings==2.9.1",
]

[tool.setuptools.packages.find]
include = ["sls_shared*"]
